+++ apps/docs/content/components/navigation/Tabs.mdx
@@ -50,25 +50,19 @@ Tabs items can be rendered dynamically.
 
 By default, tabs are activated automatically. This means when you use the arrow keys to change tabs, the tab is activated and focused.
 
-```
+<Callout variant="information">
 The content of a tab should ideally be preloaded. However, if switching to a tab panel causes a network request and possibly a page refresh, there might be some notable latency and this might affect the experience for keyboard and screen reader users.
-```
+</Callout>
 
 In this scenario, you should use a manually activated tab, which moves focus without activating the tabs. With the focus on a specific tab, users can activate a tab by pressing Space or Enter.
 
 <Example src="Tabs/docs/manual" />
 
-### In card
+### Variants
 
-A tabs component can be rendered as an in-card variant.
+Tabs can use different variants.
 
-<Example src="Tabs/docs/in-card" />
-
-### Heading
-
-A tabs component can be rendered as a heading variant.
-
-<Example src="Tabs/docs/heading" />
+<Example src="Tabs/docs/variant" />
 
 ### Fluid
