+++ packages/components/src/Tabs/docs/migration-notes.md
@@ -2,4 +2,4 @@ Coming from Orbiter, you should be aware of the following changes:
 
 - `orientation` has been removed, not supported.
 - `disabled` has been removed, not supported.
-- `manual` has been removed. Refer to this [sample](#manual) to quickly match old sizes.
+- `manual` has been removed. Refer to this [sample](#usage-manually-activated-tabs) to quickly match old sizes.
