+++ packages/components/src/Tabs/src/TabList.tsx
@@ -18,7 +18,7 @@ export interface TabListProps<T> extends
 
 function TabList<T extends object>(props: TabListProps<T>, ref: ForwardedRef<HTMLDivElement>) {
     [props, ref] = useContextProps(props, ref, TabListContext);
-    const { variant, ...lineProps } = useContext(InternalTabsContext) ?? {};
+    const { variant } = useContext(InternalTabsContext) ?? {};
     const state = useContext(TabListStateContext);
 
     const [selectedTab, setSelectedTab] = useState<HTMLElement | undefined>(undefined);
@@ -67,7 +67,7 @@ function TabList<T extends object>(props: TabListProps<T>, ref: ForwardedRef<HTM
             >
                 {children}
             </RACTablist>
-            <TabLine {...lineProps} selectedTab={selectedTab} />
+            <TabLine selectedTab={selectedTab} />
         </div>
     );
 }
